+++ src/include/database/database.h
@@ -22,5 +22,5 @@ public:
   virtual int initialize() = 0;
   virtual int create_user(user) = 0;
   virtual std::vector<std::string> list_users() = 0;
-  virtual int count_user() = 0;
+  virtual int64_t count_user() = 0;
 };
